search icon refactor
