+++ src/App/Playlist/LoadedPlaylist/Sidebar/LockIcon/Lock.scss
