+++ src/App/Playlist/LoadedPlaylist/Sidebar/LockIcon/index.js
