+++ src/App/Playlist/LoadedPlaylist/Sidebar/SearchIcon/Search.scss
@@ -1,0 +1,16 @@
+@import 'styles/_transitions.scss';
+
+.search-icon-container {
+  position: absolute;
+  top: 20px;
+  .search-icon {
+    width: 30px;
+    height: 30px;
+    fill: #4d4d4d;
+    cursor: pointer;
+    transition: $track-transition;
+    &:hover {
+      transform: scale(1.1);
+    }
+  }
+}
