+++ src/App/Playlist/LoadedPlaylist/Sidebar/ShuffleIcon/Shuffle.scss
