+++ src/App/Playlist/LoadedPlaylist/Sidebar/ShuffleIcon/index.js
