+++ src/App/Playlist/LoadedPlaylist/Sidebar/Sidebar.scss
@@ -1,5 +1,3 @@
-@import 'styles/_transitions.scss';
-
 .sidebar {
   position: fixed;
   height: 100vh;
@@ -9,19 +7,4 @@
   display: flex;
   justify-content: center;
   align-items: center;
-
-  .search-icon-container {
-    position: absolute;
-    top: 20px;
-    .search-icon {
-      width: 30px;
-      height: 30px;
-      fill: #4d4d4d;
-      cursor: pointer;
-      transition: $track-transition;
-      &:hover {
-        transform: scale(1.1);
-      }
-    }
-  }
 }
